requirements.txt
loguru
websocket-client
pyfiglet
colorama
pysocks
